+++ example/ios/AppSpectorPluginTests/AppSpectorPluginTests.m
@@ -16,7 +16,8 @@
 @interface AppSpectorPluginTests : XCTestCase
 
 @property (strong, nonatomic) id validatorMock;
-@property (strong, nonatomic) AppSpectorPlugin *plugin;
+@property (strong, nonatomic) AppSpectorPlugin *handler;
+@property (strong, nonatomic) FlutterMethodChannel *channel;
 
 @end
 
@@ -24,7 +25,8 @@
 
 - (void)setUp {
     self.validatorMock = OCMClassMock([ASPluginCallValidator class]);
-    self.plugin = [[AppSpectorPlugin alloc] initWithCallValidator:self.validatorMock];
+    self.channel = OCMClassMock([FlutterMethodChannel class]);
+    self.handler = [[AppSpectorPlugin alloc] initWithCallValidator:self.validatorMock channel:self.channel];
 }
 
 - (void)tearDown {
@@ -35,7 +37,10 @@
 - (void)testHandlerSupportsRunCall {
     XCTestExpectation *e = [self expectationWithDescription:@""];
     
-    FlutterMethodCall *call = [FlutterMethodCall methodCallWithMethodName:@"run" arguments:@{ @"apiKey" : @"DEADBEEF" }];
+    FlutterMethodCall *call = [FlutterMethodCall methodCallWithMethodName:@"run" arguments:@{ @"apiKey" : @"DEADBEEF",
+                                                                                              @"enabledMonitors" : @[],
+                                                                                              @"metadata" : @{}
+    }];
     OCMStub([self.validatorMock controlMethodSupported:[OCMArg any]]).andReturn(YES);
     OCMStub([self.validatorMock argumentsValid:call.arguments call:call.method error:[OCMArg anyObjectRef]]).andReturn(YES);
 
@@ -43,6 +48,88 @@
     OCMExpect(ClassMethod([sdkMock runWithConfig:[OCMArg any]]));
     
     [self.plugin handleMethodCall:call result:^(id result) {
+        expect(result).equal(@"Ok");
+        OCMVerifyAll(sdkMock);
+        [e fulfill];
+    }];
+    
+    [self waitForExpectations:@[e] timeout:0.1];
+}
+
+- (void)testHandlerSupportsStopCall {
+    XCTestExpectation *e = [self expectationWithDescription:@""];
+    
+    FlutterMethodCall *call = [FlutterMethodCall methodCallWithMethodName:@"stop" arguments:@{}];
+    OCMStub([self.validatorMock controlMethodSupported:[OCMArg any]]).andReturn(YES);
+    OCMStub([self.validatorMock argumentsValid:call.arguments call:call.method errorMessage:[OCMArg anyObjectRef]]).andReturn(YES);
+
+    id sdkMock = OCMClassMock([AppSpector class]);
+    OCMExpect(ClassMethod([sdkMock stop]));
+    
+    [self.handler handleMethodCall:call result:^(id result) {
+        expect(result).equal(@"Ok");
+        OCMVerifyAll(sdkMock);
+        [e fulfill];
+    }];
+    
+    [self waitForExpectations:@[e] timeout:0.1];
+}
+
+- (void)testHandlerSupportsIsStartedCall {
+    XCTestExpectation *e = [self expectationWithDescription:@""];
+    
+    FlutterMethodCall *call = [FlutterMethodCall methodCallWithMethodName:@"isStarted" arguments:@{}];
+    OCMStub([self.validatorMock controlMethodSupported:[OCMArg any]]).andReturn(YES);
+    OCMStub([self.validatorMock argumentsValid:call.arguments call:call.method errorMessage:[OCMArg anyObjectRef]]).andReturn(YES);
+
+    id sdkMock = OCMClassMock([AppSpector class]);
+    OCMStub(ClassMethod([sdkMock isRunning])).andReturn(YES);
+    
+    [self.handler handleMethodCall:call result:^(id result) {
+        expect(result).beTruthy();
+        [e fulfill];
+    }];
+    
+    [self waitForExpectations:@[e] timeout:0.1];
+}
+
+- (void)testHandlerSupportsSetMetadataCall {
+    XCTestExpectation *e = [self expectationWithDescription:@""];
+    
+    FlutterMethodCall *call = [FlutterMethodCall methodCallWithMethodName:@"setMetadata" arguments:@{ @"key" : @"userSpecifiedDeviceName",
+                                                                                              @"value" : @"device name"
+    }];
+    OCMStub([self.validatorMock controlMethodSupported:[OCMArg any]]).andReturn(YES);
+    OCMStub([self.validatorMock argumentsValid:call.arguments call:call.method errorMessage:[OCMArg anyObjectRef]]).andReturn(YES);
+
+    ASMetadata *expectedMetadata = @{AS_DEVICE_NAME_KEY:@"device name"};
+  
+    id sdkMock = OCMClassMock([AppSpector class]);
+    OCMExpect(ClassMethod([sdkMock updateMetadata: expectedMetadata]));
+    
+    [self.handler handleMethodCall:call result:^(id result) {
+        expect(result).equal(@"Ok");
+        OCMVerifyAll(sdkMock);
+        [e fulfill];
+    }];
+    
+    [self waitForExpectations:@[e] timeout:0.1];
+}
+
+- (void)testHandlerSupportsRemoveMetadataCall {
+    XCTestExpectation *e = [self expectationWithDescription:@""];
+    
+    FlutterMethodCall *call = [FlutterMethodCall methodCallWithMethodName:@"removeMetadata" arguments:@{ @"key" : @"userSpecifiedDeviceName"
+    }];
+    OCMStub([self.validatorMock controlMethodSupported:[OCMArg any]]).andReturn(YES);
+    OCMStub([self.validatorMock argumentsValid:call.arguments call:call.method errorMessage:[OCMArg anyObjectRef]]).andReturn(YES);
+
+    ASMetadata *expectedMetadata = @{};
+  
+    id sdkMock = OCMClassMock([AppSpector class]);
+    OCMExpect(ClassMethod([sdkMock updateMetadata: expectedMetadata]));
+    
+    [self.handler handleMethodCall:call result:^(id result) {
         expect(result).equal(@"Ok");
         OCMVerifyAll(sdkMock);
         [e fulfill];