+++ ios/Classes/ASPluginCallValidator.h
@@ -14,6 +14,11 @@ typedef NSString ASPluginMethodArgumentName;
 typedef NSDictionary <ASPluginMethodArgumentName *, id> ASPluginMethodArgumentsList;
 
 extern ASPluginMethodName * const kRunMethodName;
+extern ASPluginMethodName * const kStopMethodName;
+extern ASPluginMethodName * const kStartMethodName;
+extern ASPluginMethodName * const kIsStartedMethodName;
+extern ASPluginMethodName * const kSetMetadataMethodName;
+extern ASPluginMethodName * const kRemoveMetadataMethodName;
 extern ASPluginMethodName * const kHTTPRequestMethodName;
 extern ASPluginMethodName * const kHTTPResponseMethodName;
 extern ASPluginMethodName * const kLogEventMethodName;
@@ -39,7 +44,9 @@ extern ASPluginMethodArgumentName * const kMessageArgument;
 
 - (BOOL)argumentsValid:(ASPluginMethodArgumentsList *)arguments
                   call:(ASPluginMethodName *)methodName
-                 error:(NSError **)error;
+          errorMessage:(__autoreleasing NSString *_Nonnull*_Nullable)errorMessage;
+
+
 
 @end